requirements.txt
deepface==0.0.79
numpy==1.23.5
Pillow==9.4.0
Pillow==9.5.0
tqdm==4.64.1
